+++ src/test/java/de/alpharogroup/collections/set/SetFactoryTest.java
@@ -24,6 +24,8 @@
  */
 package de.alpharogroup.collections.set;
 
+import static org.testng.Assert.assertEquals;
+import static org.testng.Assert.assertNotNull;
 import static org.testng.AssertJUnit.assertTrue;
 
 import java.lang.reflect.InvocationTargetException;
@@ -32,6 +34,8 @@ import java.util.Comparator;
 import java.util.List;
 import java.util.Set;
 import java.util.SortedSet;
+import java.util.function.Supplier;
+import java.util.stream.Collectors;
 
 import org.meanbean.test.BeanTestException;
 import org.meanbean.test.BeanTester;
@@ -154,7 +158,6 @@ public class SetFactoryTest
 
 	}
 
-
 	/**
 	 * Test method for {@link SetFactory#newTreeSet(Collection, Comparator, T[])}.
 	 */
@@ -172,6 +175,28 @@ public class SetFactoryTest
 		set = SetFactory.newTreeSet(ListFactory.<String> newArrayList(), comparator, "foo", "bar",
 			"foo");
 		assertTrue(set.size() == 2);
+	}
+
+
+	/**
+	 * Test for method {@link SetFactory#newTreeSetSupplier}
+	 */
+	@Test
+	public void testNewTreeSetSupplier()
+	{
+		SortedSet<String> expected;
+		SortedSet<String> actual;
+		final List<String> list = ListFactory.newArrayList("foo", "fasel", "food", "barista",
+			"fao");
+		Comparator<String> stringComparator = (String one, String other) -> one.compareTo(other);
+		Supplier<SortedSet<String>> supplier = SetFactory.newTreeSetSupplier(stringComparator);
+		assertNotNull(supplier);
+		actual = list.stream().collect(Collectors.toCollection(supplier));
+		assertTrue(actual.size() == 5);
+		final Comparator<String> comparator = StringLengthComparator.of(SortOrder.ASCENDING);
+		final List<String> list2 = ListFactory.newArrayList("foo", "fasel");
+		expected = SetFactory.newTreeSet(list2, comparator, "food", "barista", "fao");
+		assertEquals(actual, expected);
 	}
 
 	/**